requirements.txt
alembic==1.4.3
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
marshmallow==3.9.0
passlib==1.7.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
PyYAML==5.3.1
SQLAlchemy==1.3.20
Flask-JWT-Extended==3.24.1
python-dotenv==0.15.0
